Move TS model to types folder
